requirements.txt
Flask
python-dotenv
flask-bcrypt